fn_extract_bip_reports_archrock/requirements.txt
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.1
azure-functions==1.12.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.14.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
idna==3.4
isodate==0.6.1
lxml==4.9.2
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.2
platformdirs==2.6.2
portalocker==2.6.0
pyarrow==10.0.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
xmltodict==0.13.0
zeep==4.2.1
